move MongoToAst to server folder special answer for database config
